requirements.txt
streamlit==1.28.0
langchain==0.1.0
sentence-transformers==2.2.2
chromadb==0.4.15
PyPDF2==3.0.1
python-docx==0.8.11
openai==1.3.0
transformers==4.35.0
torch==2.1.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
nltk==3.8.1
python-dotenv==1.0.0